% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageBounds.R
\name{imageBounds}
\alias{imageBounds}
\title{Returns the Bbox in form of an array of a given JSON}
\usage{
imageBounds(input)
}
\arguments{
\item{input}{An unnested JSON}
}
\description{
Returns the Bbox in form of an array of a given JSON
}

