% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteRidw.R
\name{inteRidw}
\alias{inteRidw}
\title{This function interpolates senseMap data using the IDW method.}
\usage{
inteRidw(input)
}
\arguments{
\item{input}{An unnested JSON}
}
\description{
This function interpolates senseMap data using the IDW method.
}

