% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteRidwIdp.R
\name{inteRidwIdp}
\alias{inteRidwIdp}
\title{This function interpolates senseMap data using the IDW method with adjustable idp}
\usage{
inteRidwIdp(input, x)
}
\arguments{
\item{input}{An unnested JSON}

\item{x}{number to be used as idp(inverse distance weighting power)}
}
\description{
This function interpolates senseMap data using the IDW method with adjustable idp
}

