% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteRidwIdp.R
\name{inteRidwTest}
\alias{inteRidwTest}
\title{This function interpolates senseMap data using the IDW method with adjustable idp}
\usage{
inteRidwTest(input, x)
}
\arguments{
\item{input}{An unnested JSON}

\item{x}{number to be used as idp(inverse distance weighting power)}
}
\description{
This function interpolates senseMap data using the IDW method with adjustable idp
}

