% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteRtp.R
\name{inteRtp}
\alias{inteRtp}
\title{This function interpolates senseMap data using the IDW method.}
\usage{
inteRtp(input)
}
\arguments{
\item{input}{An unnested JSON}
}
\description{
This function interpolates senseMap data using the IDW method.
}

