% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapper}
\alias{wrapper}
\title{This function serves as a wrapper function to set up CORS for OpenCPU}
\usage{
wrapper()
}
\description{
This function serves as a wrapper function to set up CORS for OpenCPU
}

